test: use test fixtures plugin to dry test utilities
