+++ domain/src/testFixtures/scala/io/github/positionpal/location/domain/GeoUtils.scala
@@ -1,0 +1,8 @@
+package io.github.positionpal.location.domain
+
+object GeoUtils:
+  val bolognaCampus: GPSLocation = GPSLocation(44.487912, 11.32885)
+  val imolaCampus: GPSLocation = GPSLocation(44.352962, 11.711285)
+  val forliCampus: GPSLocation = GPSLocation(44.219119, 12.042589)
+  val cesenaCampus: GPSLocation = GPSLocation(44.1476299926484, 12.2357184467018)
+end GeoUtils
