+++ domain/src/testFixtures/scala/io/github/positionpal/location/domain/TimeUtils.scala
@@ -1,0 +1,9 @@
+package io.github.positionpal.location.domain
+
+object TimeUtils:
+  import java.time.Instant
+
+  def now: Instant = Instant.now()
+  def inTheFuture: Instant = Instant.parse("2100-12-31T23:59:59Z")
+  def inThePast: Instant = Instant.parse("2000-01-01T00:00:00Z")
+end TimeUtils
