+++ infrastructure/src/test/scala/io/github/positionpal/location/infrastructure/ws/WebSocketTestDSL.scala
@@ -8,8 +8,9 @@ import akka.http.scaladsl.model.ws.{Message, TextMessage}
 import akka.stream.scaladsl.Sink
 import io.bullet.borer.Json
 import io.github.positionpal.entities.{GroupId, UserId}
+import io.github.positionpal.location.domain.TimeUtils.*
 import io.github.positionpal.location.domain.{DrivenEvent, DrivingEvent, GPSLocation, SampledLocation}
-import io.github.positionpal.location.infrastructure.{TimeUtils, WebSocketClient}
+import io.github.positionpal.location.infrastructure.WebSocketClient
 import io.github.positionpal.location.presentation.ModelCodecs
 
 trait WebSocketTestDSL:
@@ -18,7 +19,7 @@ trait WebSocketTestDSL:
   import scala.concurrent.ExecutionContext.Implicits.global
 
   def sample(userId: UserId, location: GPSLocation): SampledLocation =
-    SampledLocation(TimeUtils.now, userId, location)
+    SampledLocation(now, userId, location)
 
   case class WebSocketTestConfig(baseEndpoint: String, connectionTimeout: FiniteDuration)
